Cleanup test cases of yorkie_test.ts
Organize test cases according to the purpose of each test case.
