+++ test/integration/counter_test.ts
@@ -1,5 +1,6 @@
 import { assert } from 'chai';
 import { Document } from '../../src/document/document';
+import { withTwoClientsAndDocuments } from './integration_helper';
 
 describe('Counter', function () {
   it('can be increased by Counter type', function () {
@@ -28,5 +29,48 @@ describe('Counter', function () {
       });
     }, 'Unsupported type of value: boolean');
     assert.equal(`{"k1":{"age":8.5,"length":18.5}}`, doc.toSortedJSON());
+  });
+
+  it('Can handle increase operation', async function () {
+    await withTwoClientsAndDocuments(async (c1, d1, c2, d2) => {
+      d1.update((root) => {
+        root.createCounter('age', 0);
+      });
+      d1.update((root) => {
+        root['age'].increase(1).increase(2);
+        root.createCounter('length', 10);
+      });
+
+      await c1.sync();
+      await c2.sync();
+      assert.equal(d1.toSortedJSON(), d2.toSortedJSON());
+    }, this.test!.title);
+  });
+
+  it('Can handle concurrent increase operation', async function () {
+    await withTwoClientsAndDocuments(async (c1, d1, c2, d2) => {
+      d1.update((root) => {
+        root.createCounter('age', 0);
+        root.createCounter('width', 0);
+        root.createCounter('height', 0);
+      });
+      await c1.sync();
+      await c2.sync();
+      assert.equal(d1.toSortedJSON(), d2.toSortedJSON());
+
+      d1.update((root) => {
+        root['age'].increase(1).increase(2);
+        root['width'].increase(10);
+      });
+      d2.update((root) => {
+        root['age'].increase(3.14).increase(2);
+        root.createCounter('width', 2.5);
+      });
+      await c1.sync();
+      await c2.sync();
+      await c1.sync();
+
+      assert.equal(d1.toSortedJSON(), d2.toSortedJSON());
+    }, this.test!.title);
   });
 });
